Generate AUC plots to assess the performance of feature selection methods

Generate AUC plots to assess the performance of feature selection methods

Usage

get_auc_plot(dataset, imp_features_names)

Arguments

dataset

(data.frame) the dataset to be used. Last column must be the response variable.

imp_features_names

(data.frame) each column contains the names of the features names sorted from most to least important

Value

(gtable) a plot with the AUCs of the different models